requirements.txt
PySide6>=6.4.0
PyYAML>=6.0